rename(network scripts)
